+++ R/plotting-utils.R
@@ -2,6 +2,13 @@
 #'
 #' @param size base size of text
 #'
+#' @return a \code{\link[ggplot2]{theme}} object
+#'
+#' @examples
+#' library(ggplot2)
+#' ggplot(mtcars, aes(x = mpg, y = drat)) +
+#'     geom_point() +
+#'     theme_arial()
 #' @export
 theme_arial <- function(size = 8) {
     theme_arial <- ggplot2::theme_light() %+replace%
@@ -23,6 +30,14 @@ theme_arial <- function(size = 8) {
 
 #' Set x-axis vertical labels in ggplot2
 #'
+#' @return a \code{\link[ggplot2]{theme}} object that rotates x-axis labels to 90
+#'   degrees
+#'
+#' @examples
+#' library(ggplot2)
+#' ggplot(mtcars, aes(x = factor(cyl), y = mpg)) +
+#'     geom_boxplot() +
+#'     vertical_labs()
 #' @export
 vertical_labs <- function() {
     theme(axis.text.x = element_text(angle = 90, hjust = 1))