+++ R/tgppt-package.R
@@ -3,6 +3,14 @@
 #' @importFrom purrr modify_at %||%
 #' @import rvg
 #' @import officer
+#'
+#' @section Options:
+#' \describe{
+#'   \item{`tgppt.template`}{Path to a custom PowerPoint template file (.pptx).
+#'     When set, this template is used by default in [new_ppt()], [plot_gg_ppt()],
+#'     and [plot_base_ppt()] instead of the bundled template. The `template`
+#'     parameter in those functions takes precedence over this option.}
+#' }
 "_PACKAGE"
 
 # The following block is used by usethis to automatically manage